Issue #85 - Refactor Zone entity and database schema
Added a new `Location` property to the `Zone` class in the `digioz.Portal.Bo` namespace. Updated the `Location` column in the `Zone` table schema to reduce its maximum length from 255 to 50. Introduced a nullable `Location` column with a maximum length of 50 in another table schema.

Removed seed data insertion and deletion for the `Zone` table in the `20250723043739_MigrateOtherDataSeed.cs` migration. These changes streamline the `Zone` entity and its database representation.